requirements.txt
beautifulsoup4==4.6.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.6
django-ads==1.0.0
django-appconf==1.0.3
django-bootstrap4==1.0.1
django-braces==1.13.0
django-ckeditor==5.8.0
django-classy-tags==0.9.0
django-cms==3.7.0
django-crispy-forms==1.8.1
django-filer==1.6.0
django-formtools==2.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-modelcluster==5.0
django-mptt==0.10.0
django-polymorphic==2.0.3
django-sekizai==1.0.0
django-taggit==1.1.0
django-treebeard==4.3
djangocms-admin-style==1.4.0
djangocms-attributes-field==1.1.0
djangocms-bootstrap4==1.5.0
djangocms-file==2.3.0
djangocms-googlemap==1.3.0
djangocms-icon==1.4.2
djangocms-installer==1.2.0
djangocms-link==2.5.0
djangocms-picture==2.3.0
djangocms-snippet==2.2.0
djangocms-style==2.2.0
djangocms-text-ckeditor==3.8.0
djangocms-video==2.2.0
djangorestframework==3.10.3
draftjs-exporter==2.1.7
easy-thumbnails==2.6
gunicorn==20.0.4
html5lib==1.0.1
idna==2.8
Pillow==6.2.1
psycopg2==2.8.4
pycparser==2.19
pytz==2019.3
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tzlocal==2.0.0
Unidecode==1.0.23
urllib3==1.25.7
wagtail==2.7
webencodings==0.5.1
whitenoise==4.1.4
Willow==1.3
